Expose volume and take directory name param
In the new deployment process, we build pool containers using
`volumes-from`. To include the volumes from appium containers in the
pool container, we need to expose the volume in the appium image.

Additionally, product relies on the appium directory using the public
name of the appium version. We therefore take the directory name as a
parameter to have more control in naming the appium version.